[refactor] Move settings feature to single file
